zero warnings

+++ commandLine/src/optionparser.h
@@ -1447,7 +1447,7 @@ public:
 				return false; // overflow protection: don't accept number of options that doesn't fit signed int
 
 			buffer[parser.op_count] = option;
-			int idx = buffer[parser.op_count].desc->index;
+			size_t idx = buffer[parser.op_count].desc->index;
 			if (options[idx])
 				options[idx].append(buffer[parser.op_count]);
 			else
